Updated based on suggestions
