+++ packages/chakra-helpers/src/helpers.ts
@@ -1,22 +1,6 @@
-import React, {Context, createContext, useContext} from 'react';
-import {BiText} from '@react-icons/all-files/bi/BiText';
 import {ColorHues} from '@chakra-ui/react';
 import {ColorPalette, MonochromePalette} from '@apollo/space-kit/colors';
-import {SiGnubash} from '@react-icons/all-files/si/SiGnubash';
-import {SiGraphql} from '@react-icons/all-files/si/SiGraphql';
-import {SiGroovy} from '@react-icons/all-files/si/SiGroovy';
-import {SiJava} from '@react-icons/all-files/si/SiJava';
-import {SiJavascript} from '@react-icons/all-files/si/SiJavascript';
-import {SiJson} from '@react-icons/all-files/si/SiJson';
-import {SiKotlin} from '@react-icons/all-files/si/SiKotlin';
-import {SiRuby} from '@react-icons/all-files/si/SiRuby';
-import {SiRust} from '@react-icons/all-files/si/SiRust';
-import {SiSwift} from '@react-icons/all-files/si/SiSwift';
-import {SiTypescript} from '@react-icons/all-files/si/SiTypescript';
-import {TiDocumentText} from '@react-icons/all-files/ti/TiDocumentText';
-
 import {mix} from 'polished';
-import type {IconType} from '@react-icons/all-files/lib';
 
 const body = "'Source Sans Pro', sans-serif";
 export const fonts = {
@@ -65,68 +49,3 @@ export const createColorPalette = (color: ColorPalette): ColorHues => ({
   800: mix(0.5, color.darker, color.darkest),
   900: color.darkest
 });
-
-export function getNormalizedLanguage(language?: string): string {
-  const classless =
-    language
-      ?.replace(/language-/g, '')
-      .toLocaleLowerCase()
-      .replace(/:.*/g, '') ?? '';
-  switch (classless) {
-    case 'js':
-    case 'jsx':
-    case 'javascript':
-      return 'JavaScript';
-    case 'ts':
-    case 'tsx':
-    case 'typescript':
-      return 'TypeScript';
-    case 'graphql':
-      return 'GraphQL';
-    case 'json':
-      return 'JSON';
-    case 'yaml':
-      return 'YAML';
-    case 'text':
-    case 'swift':
-    case 'bash':
-    case 'groovy':
-    case 'java':
-    case 'kotlin':
-    case 'rust':
-    case 'ruby':
-      return classless[0].toLocaleUpperCase() + classless.slice(1);
-    default:
-      return classless;
-  }
-}
-
-export function getIconComponentType(language: string): IconType | undefined {
-  const normalizedLang = getNormalizedLanguage(language);
-  const componentMap: Record<string, IconType> = {
-    JavaScript: SiJavascript,
-    TypeScript: SiTypescript,
-    JSON: SiJson,
-    GraphQL: SiGraphql,
-    Text: BiText,
-    Bash: SiGnubash,
-    Groovy: SiGroovy,
-    Java: SiJava,
-    Kotlin: SiKotlin,
-    Ruby: SiRuby,
-    Rust: SiRust,
-    Swift: SiSwift,
-    YAML: TiDocumentText
-  };
-  const component = componentMap[normalizedLang];
-  return component;
-}
-
-export function getIconComponent(language: string): React.ReactElement {
-  const normalizedLang = getNormalizedLanguage(language);
-  const iconType = getIconComponentType(normalizedLang);
-  if (!iconType) {
-    return undefined;
-  }
-  return React.createElement(iconType);
-}
